refactor(extract base class):
